Read fraction formats correctly from ODF 1.2 files.
2011-06-09  Andreas J. Guelzow <aguelzow@pyrshep.ca>

	* developer/odf-foreign.txt: update

2011-06-09  Andreas J. Guelzow <aguelzow@pyrshep.ca>

	* openoffice-read.c (odf_fraction): handle absence of min_i_digits
	correctly depending on ODF version
	(odf_find_version): new
	(styles_dtd): connect odf_find_version and add a bunch of stubs
	(opendoc_content_dtd): connect odf_find_version
	(openoffice_file_open): initialize new field ver_odf

+++ src/gui-util.c
@@ -769,7 +769,7 @@ gnm_create_std_tags_for_buffer (GtkTextBuffer *buffer)
 static gint
 gnm_load_pango_byte_to_char (gchar const *str, gint byte)
 {
-	if (byte >= strlen (str))
+	if (byte >= (gint) strlen (str))
 		return g_utf8_strlen (str, -1);
 	return g_utf8_pointer_to_offset (str, 
 					 g_utf8_prev_char (str + byte + 1));
